fix: use MemoryCard to save clientId as session
