+++ src/config.ts
@@ -11,7 +11,7 @@ export {
 
 const VERSION = packageJson.version || '0.0.0'
 
-const MEMORY_SLOT = 'PUPPET_WHATSAPP'
+const MEMORY_SLOT = 'PUPPET_WHATSAPP_CLIENT_ID'
 const PRE = 'PuppetWhatsApp'
 
 export const SPECIAL_BOT_PUSHNAME = '-' // FIXME: pushname is '-', see: https://github.com/wechaty/puppet-whatsapp/issues/233
